\name{AddXMLInternal}
\title{Internal functions for adding the necessary XML content for accessible graphs in SVG format}
\description{Mostly internal functions for adding XML content to an external file}
\usage{
}
\arguments{
 \item{x}{an object that is of the form needed by the internal function concerned.}
\item{file}{the XML file to be written to.}
}
\details{
\value{Either the filename written to, or NULL}
\author{Volker Sorge and A. Jonathan R. Godfrey}
