\name{BrailleR-package}
\alias{BrailleR-package}
\alias{BrailleR}
\docType{package}
\title{
Improved Access for Blind Users}
\description{Blind users do not have access to the graphical output from R without printing the content of graph windows to an embosser of some kind. This is not as immediate as is required for efficient access to statistical output. The functions here are created so that blind people can make even better use of R. This includes the text descriptions of graphs, convenience functions to replace the functionality offered in many GUI front ends, and experimental functionality for optimising graphical content to prepare it for embossing as tactile images.
}
\details{
\tabular{ll}{
Package: \tab BrailleR\cr
Type: \tab Package\cr
Version: \tab 0.25.7\cr
Date: \tab 2017-04-30\cr
License: \tab GPL-2\cr
}
}
\author{
A. Jonathan R. Godfrey Volker Sorge and Timothy P. Bilton with other contributions.

Maintainer: A. Jonathan R. Godfrey <a.j.godfrey@massey.ac.nz>
}
\references{
Godfrey, A.J.R. (2013) \dQuote{Statistical Software from a Blind Person's Perspective: R is the Best, but we can make it better}, The R Journal 5(1), pp73-79.
}
\keyword{ package }
