\name{BrailleRHome}
\alias{BrailleRHome}
\title{Open the BrailleR Project home page in your browser}
\usage{
BrailleRHome()
}
\value{
Nothing. The functionis for opening a web page and nothing will result in the R session.
}
\description{
Visit the BrailleR Project home page for updates, instructions on how to join the mailing list which we call BlindRUG and get the latest downloads.
}
\author{
A. Jonathan R. Godfrey
}

