\name{Deprecated}
\alias{GetPython27}
\alias{GetWxPython27}
\alias{MakeSlideShow}
\alias{MakeSlidy}
\alias{MakeAllInOneSlide}
\title{Deprecated functions}
\usage{
GetPython27(...) 
GetWxPython27(...) 

MakeAllInOneSlide(...)
MakeSlidy(...)
akeSlideShow(...)
}
\arguments{
\item{...}{cover for unnecessary arguments going to deprocated functions.}
}
\details{ These functions are redundant.


A message is printed to inform the user if they are called.}
\value{
NULL.
}
\description{
Functions that have become redundant courtesy of development outside the BrailleR package.
}
\author{
A. Jonathan R. Godfrey, is responsible for the deprecation.
}

