\name{GetWriteR}
\alias{GetWriteR}
\title{Download an executable for Windows users}
\usage{
GetWriteR()
}
\value{
NULL. The downloaded file is saved in the user's MyBrailleR folder.
}
\description{
Users that do not have Pyton 2.7 and wxPython installed cannot use the WriteR application file as provided by the BrailleR package. An executabel for Windows is available for separate download. The downloaded file is saved into the user's MyBrailleR folder.
}
\author{
A. Jonathan R. Godfrey
}

