\name{Internal}
\alias{GetAxisTicks}
\alias{InQuotes}
\title{Internal functions for the BrailleR package}
\description{Some functions that probably weren't really necessary, but proved useful at some stage.}
\usage{
GetAxisTicks(x)

InQuotes(x)
}
\arguments{
  \item{x}{
an object that is of the form needed by the internal function concerned.}
}
\details{
\code{GetAxisTicks} is intended to create the sequence of values used on an axis.}
\value{These should be fairly obvious from the name of the function}
\author{A. Jonathan R. Godfrey}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
