\name{MakeSlideShow}
\title{Turn a set of Rmd files into an HTML slide show}
\description{
These functions used to take a set of Rmd files in alphabetical order and makes them a set of linked HTML files or a single slidy presentation suitable for delivering a presentation., or a single plain HTML file suitable for distribution They have been overtaken by bookdown, pkgdown, rmarkdown, etc.}

