\name{AddXMLInternal}
\alias{.AddXMLAddTitle}
\title{Internal functions for adding the necessary XML content for accessible graphs in SVG format}
\description{Mostly internal functions for adding XML content to an external file}
\details{Not for use as exported functions}
\value{Either the filename written to, or NULL}
\author{Volker Sorge and A. Jonathan R. Godfrey}
