\name{BrailleR-package}
\alias{BrailleR-package}
\alias{BrailleR}
\docType{package}
\title{
Improved access for blind useRs}
\description{Blind useRs do not have access to the graphical output from R without printing the content of graph windows to an embosser of some kind. This is not as immediate as is required for efficient access to statistical output. 

Various functions are also being developed to make text output (that is visually appealing) more useful for a blind useR who is reliant on synthesized speech or braille output to interpret the results.

Ultimately, the functions here are created so that blind people can make even better use of R.
}
\details{
\tabular{ll}{
Package: \tab BrailleR\cr
Type: \tab Package\cr
Version: \tab 0.8\cr
Date: \tab 2013-08-27\cr
License: \tab GPL-2\cr
}
}
\author{
Jonathan Godfrey

Maintainer: Jonathan Godfrey <a.j.godfrey@massey.ac.nz>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
