\name{BrailleR-package}
\alias{BrailleR-package}
\alias{BrailleR}
\docType{package}
\title{
Improved access for blind useRs}
\description{Blind useRs do not have access to the graphical output from R without printing the content of graph windows to an embosser of some kind. This is not as immediate as is required for efficient access to statistical output. 

Various functions are also being developed to make text output (that is visually appealing) more useful for a blind useR who is reliant on synthesized speech or braille output to interpret the results.

Ultimately, the functions here are created so that blind people can make even better use of R.
}
\details{
\tabular{ll}{
Package: \tab BrailleR\cr
Type: \tab Package\cr
Version: \tab 0.11\cr
Date: \tab 2014-09-27\cr
License: \tab GPL-2\cr
}
}
\author{
A. Jonathan R. Godfrey

Maintainer: A. Jonathan R. Godfrey <a.j.godfrey@massey.ac.nz>
}
\references{
Godfrey, A.J.R. (2013) \sQuote{Statistical Software from a Blind Person's Perspective: R is the Best, but we can make it better}, The R Journal 5(1), pp73-80.
}
\keyword{ package }
