\name{BrailleRUsefulLinks}
\alias{BrailleRHome}
\alias{LURN}
\title{Open the BrailleR Project or Let's Use R Now home page in your browser}
\usage{
BrailleRHome()
LURN(BlindVersion = getOption("BrailleR.VI"))
}
\arguments{
\item{BlindVersion}{Use the version of Let's Use R Now tailored to an audience of blind users.}
}
\value{
Nothing. These functions are for opening a web page and nothing will result in the R session.
}
\description{
Visit the BrailleR Project home page for updates, instructions on how to join the mailing list which we call BlindRUG and get the latest downloads; or the Let's Use R Now (LURN) home page to read an online manual for using R.
}
\author{
A. Jonathan R. Godfrey
}

