
\name{SetupBrailleR}
\alias{SetupBrailleR}
\title{Establish the BrailleR folder for the user}
\usage{
SetupBrailleR()
}
\value{
The path to the folder is returned invisibly.
}
\description{
Creates a permanent folder if the user agrees.
}
\details{
The user can establish a permanent folder if the BrailleR package is loaded in an interactive session. If only used in batch mode, the user will only be able to use the default BrailleR settings.
}
\author{
A. Jonathan R. Godfrey with suggestions from Henrik Bengtsson and Brian Ripley.
}

